- Feat: general refactoring - Feat: separate busineess logic form UI, add custom hooks - Feat: separate navbar content into smaller components
