+++ src/app/create-prompt/constants/index.ts
@@ -8,3 +8,6 @@ export const defaultPostPrompt: PostPromptType = {
   prompt: "",
   tag: "",
 };
+
+export const HEADER: string = `${FormType.Create} Post`
+export const DESC: string = `${FormType.Create} and share your prompt and inspire the new AI generation`
