+++ src/app/create-prompt/page.tsx
@@ -1,15 +1,17 @@
 "use client";
 
 import Form from "@components/Form/Form";
-import { useState } from "react";
-import { PostPromptType } from "@types";
-import { defaultPostPrompt, FormType } from "@app/create-prompt/constants";
+import {
+  defaultPostPrompt,
+  DESC,
+  FormType,
+  HEADER,
+} from "@app/create-prompt/constants";
+import useCreatePrompt from "@app/create-prompt/hooks/useCreatePrompt";
 
 const CreatePrompt = () => {
-  const [post, setPost] = useState<PostPromptType>(defaultPostPrompt);
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-
-  const submitPrompt = () => {};
+  const { post, setPost, isSubmitting, setIsSubmitting, submitPrompt } =
+    useCreatePrompt(defaultPostPrompt);
 
   return (
     <div>
@@ -19,6 +21,8 @@ const CreatePrompt = () => {
         setPost={setPost}
         isSubmitting={isSubmitting}
         handleSubmit={submitPrompt}
+        header={HEADER}
+        desc={DESC}
       />
     </div>
   );
