+++ src/components/Navbar/hooks/useDropdown.ts
@@ -1,0 +1,10 @@
+import { useState } from "react";
+
+export default function useDropdown() {
+  const [showDropdown, setShowDropdown] = useState<boolean>(false);
+
+  return {
+    showDropdown,
+    setShowDropdown,
+  };
+}
